Refactor code to use custom hook for cart management
The commit includes a refactoring of the cart-related functionalities out of the App and Header components, into a custom hook named 'useCart'. This hook now houses the state and logic around cart management, making the code more modular and maintainable. The components now get their cart-related props from the custom hook, helping to clean reader code, improve readability and ease future updates.
